redo mac opts
